refactor(crypto): extract transaction verification out of the model
