+++ __tests__/unit/core-transactions/handler.test.ts
@@ -635,7 +635,7 @@ describe.skip("MultiSignatureRegistrationTransaction", () => {
             delete wallet.multisignature;
 
             wallet.verifySignatures = jest.fn(() => true);
-            Transactions.Transaction.verifySecondSignature = jest.fn(() => true);
+            Transactions.TransactionVerifier.verifySecondSignature = jest.fn(() => true);
 
             instance.data.asset.multisignature.keysgroup.splice(0, 5);
             expect(() => handler.canBeApplied(instance, wallet)).toThrow(MultiSignatureMinimumKeysError);
@@ -645,7 +645,7 @@ describe.skip("MultiSignatureRegistrationTransaction", () => {
             delete wallet.multisignature;
 
             wallet.verifySignatures = jest.fn(() => true);
-            Transactions.Transaction.verifySecondSignature = jest.fn(() => true);
+            Transactions.TransactionVerifier.verifySecondSignature = jest.fn(() => true);
 
             instance.data.signatures.splice(0, 5);
             expect(() => handler.canBeApplied(instance, wallet)).toThrow(MultiSignatureKeyCountMismatchError);
