+++ packages/core-jest-matchers/src/transactions/valid.ts
@@ -16,7 +16,7 @@ expect.extend({
         let verified: boolean = false;
 
         try {
-            verified = Transactions.Transaction.verifyData(actual);
+            verified = Transactions.TransactionVerifier.verifyHash(actual);
         } catch (e) {} // tslint:disable-line
 
         return {
