+++ packages/core-transactions/src/handlers/transaction.ts
@@ -44,7 +44,7 @@ export abstract class TransactionHandler implements ITransactionHandler {
         }
 
         if (wallet.secondPublicKey) {
-            if (!Transactions.Transaction.verifySecondSignature(data, wallet.secondPublicKey)) {
+            if (!Transactions.TransactionVerifier.verifySecondSignature(data, wallet.secondPublicKey)) {
                 throw new InvalidSecondSignatureError();
             }
         } else {
