+++ packages/crypto/src/interfaces/transactions.ts
@@ -17,7 +17,7 @@ export interface ITransaction {
     deserialize(buf: ByteBuffer): void;
 
     verify(): boolean;
-    validateSchema(strict?: boolean): ISchemaValidationResult;
+    verifySchema(strict?: boolean): ISchemaValidationResult;
 
     toJson(): ITransactionJson;
