+++ packages/crypto/src/transactions/index.ts
@@ -3,5 +3,6 @@ export * from "./deserializer";
 export * from "./factory";
 export * from "./serializer";
 export * from "./types";
+export * from "./verifier";
 
 export { transactionRegistry as TransactionRegistry, TransactionConstructor } from "./registry";
